feat: Adds optional support for giving a complete URL as the ID field (#22)
With this feature, it is now supported to pass the complete URL as the `id` field and this application will do a best-effort attempt at inferring the ID from the URL.

This might not work for all URLs, as the ID can be stored in different ways. If the ID cannot be parsed from the URL, please fall back to manually parsing the ID field.